🚚 moved dtos in the respective package
